+++ lib/index.js
@@ -1,2 +1,2 @@
-const MakeResume = require("./models/jsonToCV");
+const MakeResume = require("./models/makeResume");
 module.exports = MakeResume;
